Transform workspace to module Move logic and UI into module, Left only page in pages, import all parts from module
