+++ src/modules/workspace/index.ts
@@ -1,0 +1,2 @@
+export { useWorkspace } from "./useWorkspace.hook";
+export { Workspace } from "./workspace.ui.component";
